+++ dom/media/gmp/GMPUtils.h
@@ -15,7 +15,6 @@
 #include "nsClassHashtable.h"
 
 #define CHROMIUM_CDM_API_BACKWARD_COMPAT "chromium-cdm9-host4"
-// TODO: The following should be used as we switch to cdm10
 #define CHROMIUM_CDM_API "chromium-cdm10-host4"
 
 class nsIFile;